Trace nRF54 BM reconnect in security manager
